+++ app/src/main/java/com/codesoom/assignment/utils/TaskIdGenerator.java
@@ -1,18 +1,6 @@
 package com.codesoom.assignment.utils;
 
-import org.springframework.stereotype.Component;
+public interface TaskIdGenerator {
 
-import java.util.concurrent.atomic.AtomicLong;
-
-@Component
-public class TaskIdGenerator {
-
-    private final AtomicLong atomicLong = new AtomicLong(1L);
-
-    private TaskIdGenerator() {
-    }
-
-    public Long allocateId() {
-        return atomicLong.getAndIncrement();
-    }
+    Long allocateId();
 }
